% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{calculateRMST}
\alias{calculateRMST}
\title{Calculate counterfactuals from simulated data}
\usage{
calculateRMST(coarsenedTime, survCurve)
}
\arguments{
\item{coarsenedTime}{Output from coarsenData}

\item{survCurve}{Output from calculateSurvCurve}
}
\description{
Calculate counterfactuals from simulated data
}
