% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{calculateSurvCurve}
\alias{calculateSurvCurve}
\title{Calculate counterfactuals from simulated data}
\usage{
calculateSurvCurve(coarsenedTime, survHaz)
}
\arguments{
\item{coarsenedTime}{Output from coarsenData}

\item{survHaz}{Output from estimateHaz}
}
\description{
Calculate counterfactuals from simulated data
}
