% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{coarsenData}
\alias{coarsenData}
\title{Coarsen data to non-uniform time intervals}
\usage{
coarsenData(time, outcome, nInt = NULL)
}
\arguments{
\item{nInt}{number of time intervals for coarsening the data}
}
\description{
Coarsen data to non-uniform time intervals
}
