% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogistic.R
\name{coef_pooled}
\alias{coef_pooled}
\title{Coefficients for pooled logistic regression (via iterative reweighted least square)}
\usage{
coef_pooled(
  X_baseline,
  is.temporal,
  temporal_effect,
  timeEffect,
  time,
  eventObserved,
  estimate_hazard,
  lambda,
  maxiter,
  threshold,
  printIter,
  initial_coef
)
}
\arguments{
\item{X_baseline}{Baseline variables that won't interact with time in regression,
sparse matrix of class "dgTMatrix".
Rows are ordered into decreasing survival time.}

\item{is.temporal}{Whether there is temporal effect, i.e. whether time t is a variable in the regression}

\item{temporal_effect}{Baseline variables that will interact with time in regression,
sparse matrix of class "dgTMatrix" or matrix.
Rows are ordered into decreasing survival time}

\item{timeEffect}{Functions of time in the discrete censoring hazards model.
Options currently include "linear", "ns", NULL (if is.temporal = FALSE)}

\item{time}{Observed survival time. Ordered into decreasing observed survival time}

\item{eventObserved}{Event indicator. Ordered into decreasing observed survival time}

\item{estimate_hazard}{"survival" or "censoring"}

\item{lambda}{Penalized parameter for ridge regression, a scalar. If lambda = NULL, then no penalization.}

\item{maxiter}{Maximum iterations}

\item{threshold}{Threshold for convergence}

\item{printIter}{TRUE/FALSE. Whether to print iterations or not}
}
\value{
A vector of coefficients in the order of: intercept, baseline covariates, time,
interaction term between baseline covariates and time.
And the standard error of the estimated coefficients
}
\description{
Coefficients for pooled logistic regression (via iterative reweighted least square)
}
