% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSparse.R
\name{coefSparse}
\alias{coefSparse}
\title{Coefficients for logistic regression with sparse matrix (via iterative reweighted least square)}
\usage{
coefSparse(outcome, offset, H, maxiter, threshold, initial_coef, printIter)
}
\description{
Coefficients for logistic regression with sparse matrix (via iterative reweighted least square)
}
