% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{computeEstimator}
\alias{computeEstimator}
\title{Run the algorithm}
\usage{
computeEstimator(
  treatment,
  outcome,
  time,
  initial_survHaz,
  cenHaz,
  treatProb,
  coarseningParam,
  estimand = "both",
  method,
  hazMethod = NULL
)
}
\description{
Run the algorithm
}
