% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{crossFit}
\alias{crossFit}
\title{Index for cross-fitting}
\usage{
crossFit(eventObserved, id, crossFitNum)
}
\arguments{
\item{crossFitNum}{For cross-fitting: random partition of subjects into J prediction sets of approximately the same size.}
}
\value{

}
\description{
Index for cross-fitting
}
