% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditiveHazard.R
\name{estimateAdditiveHazard}
\alias{estimateAdditiveHazard}
\title{Estimate average treatment effect under additive hazard model}
\usage{
estimateAdditiveHazard(
  eventTime,
  censorTime,
  treatment,
  pscore,
  pscoreAdjustment = "spline",
  adjustmentSpec = list(splineDf = 5, numStrata = 5),
  timeVariedPscoreEffect = FALSE,
  timeVariedTreatmentEffect = FALSE
)
}
\arguments{
\item{pscoreAdjustment}{Takes a value of either 'stratification' or 'spline'.
'match' may be supported in the future.}

\item{timeVariedPscoreEffect}{Time-varying covariates effects collapse into
time-varying propensity score effects. In other words, set this parameter
to TRUE if you want the effects of the original confounders to be time-varying.}
}
\description{
Estimate average treatment effect under additive hazard model
}
