% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_RMST.R
\name{estimateAIPWrmst}
\alias{estimateAIPWrmst}
\title{Estimate cross-fitted Augmented IPW of restricted mean survival time (RMST)}
\usage{
estimateAIPWrmst(dlong, survHaz, cenHaz, treatProb, tau)
}
\arguments{
\item{dlong}{Long-format survival data from function transformData(dwide, freqTime), must include columns: id, t, treatment, Lt, It}

\item{survHaz}{Data frame with two columns: SurvHaz1, SurvHaz0
Estimated survival hazards for each person at each time points if receive treatment 1 (SurvHaz1) and if receive treatment 0 (SurvHaz0)}

\item{cenHaz}{Data frame with two columns: CenHaz1, CenHaz0
Estimated censoring hazards for each person at each time points if receive treatment 1 (CenHaz1) and if receive treatment 0 (CenHaz0)}

\item{treatProb}{Estimated probability for treatment for each person if receive treatment 1}

\item{tau}{Time of interest. Can be a vector (multiple time of interest)}
}
\value{
A data frame with three columns: RMST1, RMST0, SErmstDiff
}
\description{
Estimate cross-fitted Augmented IPW of restricted mean survival time (RMST)
}
