% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProb.R
\name{estimateCenHaz}
\alias{estimateCenHaz}
\title{Estimate discrete censoring hazards}
\usage{
estimateCenHaz(
  dlong,
  covariates,
  covIdCenHaz,
  crossFitNum = 1,
  index_ls = NULL,
  timeEffect,
  interactWithTime,
  cenHazEstimate
)
}
\arguments{
\item{dlong}{Long-format survival data from function transformData(dwide, freqTime)}

\item{covIdCenHaz}{Covariates id to include in modeling discrete censoring hazards. If NULL, then include all covariates in the model}

\item{crossFitNum}{For cross-fitting: random partition of subjects into XXX prediction sets of approximately the same size.
If crossFitNum = 1, then no cross-fitting (default)}

\item{index_ls}{Index for cross-fitting
Default is no cross-fitting, index_ls = NULL}

\item{timeEffect}{Functions of time in the discrete censoring hazards model.
Options currently include "linear", "ns"}

\item{interactWithTime}{Data frame that include variables that interact with time in the censoring hazards model}

\item{cenHazEstimate}{Model for estimating censoring hazards. Options currently include "LASSO", "glm"}
}
\value{
A data frame with columns: ID, CenHaz1, CenHaz0
}
\description{
Estimate discrete censoring hazards
}
