% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProb.R
\name{estimateHaz}
\alias{estimateHaz}
\title{Estimate discrete survival/censoring hazards}
\usage{
estimateHaz(
  id,
  treatment,
  eventObserved,
  time,
  offset_t,
  offset_X = FALSE,
  weight = NULL,
  covariates,
  covIdHaz,
  sigma = exp(seq(log(1), log(0.01), length.out = 20)),
  crossFitNum = 1,
  index_ls = NULL,
  breaks,
  nInt = NULL,
  timeEffect,
  evenKnot,
  penalizeTimeTreatment = NULL,
  interactWithTime,
  hazEstimate,
  intercept = TRUE,
  estimate_hazard,
  getHaz,
  coef_H,
  robust = FALSE,
  threshold = 1e-08
)
}
\arguments{
\item{covIdHaz}{Covariates id to include in modeling discrete hazards. If NULL, then include all covariates in the model}

\item{crossFitNum}{For cross-fitting: random partition of subjects into XXX prediction sets of approximately the same size.
If crossFitNum = 1 and index_ls = NULL, then no cross-fitting (default)
If crossFitNum = 1/2/3/4/5 and index_ls is not NULL, then perform crossFitNum'th fold cross-fit}

\item{index_ls}{Index for cross-fitting
Default is no cross-fitting, index_ls = NULL}

\item{timeEffect}{Functions of time in the discrete hazards model.
Options currently include "linear", "ns"}

\item{interactWithTime}{Data frame that include variables that interact with time in the hazards model}

\item{hazEstimate}{Model for estimating censoring hazards. Options currently include "glm", "ridge".
If hazEstimate = NULL, then must provide coef_Haz to have prediction of hazards}

\item{estimate_hazard}{"survival" or "censoring"}

\item{coef_H}{Supply output from pooledLogistic to have prediction of hazards}

\item{maxTimePredict}{Maximum time for prediction}
}
\value{
A data frame with columns: ID, Haz1, Haz0
}
\description{
Estimate discrete survival/censoring hazards
}
