% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProb.R
\name{estimateHazards}
\alias{estimateHazards}
\title{Get estimated hazards}
\usage{
estimateHazards(
  coarsenedData,
  outcome,
  treatment,
  covariates,
  covId,
  nInt = NULL,
  hazEstimate,
  hazMethod
)
}
\arguments{
\item{nInt}{number of time intervals for coarsening the data}

\item{hazEstimate}{"survival" or "censoring"}

\item{hazMethod}{"twoStage" or "ns"}

\item{seed}{to set.seed()}
}
\description{
Get estimated hazards
}
