% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPW.R
\name{estimateIPW}
\alias{estimateIPW}
\title{Estimate IPW of survival probability / rmst at time tau}
\usage{
estimateIPW(
  treatment,
  eventObserved,
  time,
  cenHaz,
  treatProb,
  tau,
  timeIntMidPoint,
  timeIntLength,
  estimand,
  printTau
)
}
\arguments{
\item{cenHaz}{Data frame with two columns: Haz1, Haz0
Estimated censoring hazards for each person at each time points if receive treatment 1 (Haz1) and if receive treatment 0 (Haz0)}

\item{treatProb}{Estimated probability for treatment for each person if receive treatment 1}

\item{tau}{Time of interest. Can be a vector (multiple time of interest)}
}
\value{
A data frame with three columns: estimand1, estimand0, SE
}
\description{
Estimate IPW of survival probability / rmst at time tau
}
