% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProb.R
\name{estimateSurvHaz}
\alias{estimateSurvHaz}
\title{Estimate cross-fitted discrete survival hazards}
\usage{
estimateSurvHaz(
  dlong,
  covariates,
  covIdSurvHaz,
  crossFitNum = 1,
  index_ls = NULL,
  timeEffect,
  interactWithTime,
  survHazEstimate
)
}
\arguments{
\item{dlong}{Long-format survival data from function transformData(dwide, freqTime)}

\item{covIdSurvHaz}{Covariates id to include in modeling discrete survival hazards. If NULL, then include all covariates in the model}

\item{crossFitNum}{For cross-fitting: random partition of subjects into XXX prediction sets of approximately the same size.
If crossFitNum = 1, then no cross-fitting (default)}

\item{index_ls}{Index for cross-fitting
Default is no cross-fitting, index_ls = NULL}

\item{interactWithTime}{Data frame that include variables that interact with time in the censoring hazards model}

\item{survHazEstimate}{Model for estimating survival hazards. Options currently include logistic LASSO, glm}
}
\value{
A data frame with columns: ID, SurvHaz1, SurvHaz0
}
\description{
Estimate cross-fitted discrete survival hazards
}
