% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_S.R
\name{estimateTMLEprob}
\alias{estimateTMLEprob}
\title{Estimate cross-fitted TMLE of survival probability at time tau}
\usage{
estimateTMLEprob(dlong, survHaz, cenHaz, treatProb, tau)
}
\arguments{
\item{dlong}{Long-format survival data from function transformData(dwide, freqTime), must include columns: id, t, treatment, Lt, It}

\item{survHaz}{Data frame with two columns: SurvHaz1, SurvHaz0
Estimated survival hazards for each person at each time points if receive treatment 1 (SurvHaz1) and if receive treatment 0 (SurvHaz0)}

\item{cenHaz}{Data frame with two columns: CenHaz1, CenHaz0
Estimated censoring hazards for each person at each time points if receive treatment 1 (CenHaz1) and if receive treatment 0 (CenHaz0)}

\item{treatProb}{Estimated probability for treatment for each person if receive treatment 1}

\item{tau}{Time of interest. Can be a vector (multiple time of interest)}
}
\value{
A data frame with three columns: SurvProb1, SurvProb0, SEprobDiff
}
\description{
Estimate cross-fitted TMLE of survival probability at time tau
}
