% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_S_monotone.R
\name{estimateTMLEprob_monotone}
\alias{estimateTMLEprob_monotone}
\title{Estimate (cross-fitted) TMLE of survival probability at time 1, 2, ...tau}
\usage{
estimateTMLEprob_monotone(
  treatment,
  eventObserved,
  time,
  survHaz,
  cenHaz,
  treatProb,
  tau,
  timeIntMidPoint,
  printIter
)
}
\arguments{
\item{survHaz}{Data frame with two columns: Haz1, Haz0
Estimated survival hazards for each person at each time points if receive treatment 1 (SurvHaz1) and if receive treatment 0 (SurvHaz0)}

\item{cenHaz}{Data frame with two columns: Haz1, Haz0
Estimated censoring hazards for each person at each time points if receive treatment 1 (CenHaz1) and if receive treatment 0 (CenHaz0)}

\item{treatProb}{Estimated probability for treatment for each person if receive treatment 1}

\item{tau}{Max time of interest. A scalar}
}
\value{
A data frame with three columns: estimand1, estimand0, SE
}
\description{
Estimate (cross-fitted) TMLE of survival probability at time 1, 2, ...tau
}
