% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProb.R
\name{estimateTreatProb}
\alias{estimateTreatProb}
\title{Estimate treatment probability}
\usage{
estimateTreatProb(
  id,
  treatment,
  covariates,
  covIdTreatProb,
  treatProbEstimate,
  maxCohortSizeForFitting,
  index_ls = NULL,
  crossFitNum = 1
)
}
\arguments{
\item{covIdTreatProb}{Covariates id to include in modeling treatment probability. If NULL, then include all covariates in the model}

\item{treatProbEstimate}{Model for estimating treatment probability. Options currently include logistic LASSO}

\item{maxCohortSizeForFitting}{If the target or comparator cohort are larger than this number, they
will be downsampled before fitting the propensity model. The model
will be used to compute propensity scores for all subjects. The
purpose of the sampling is to gain speed.}

\item{index_ls}{Index for cross-fitting;
Default is no cross-fitting, index_ls = NULL}

\item{crossFitNum}{For cross-fitting: random partition of subjects into XXX prediction sets of approximately the same size.
If crossFitNum = 1, then no cross-fitting (default)}
}
\value{
A data frame with columns: id, TreatProb
}
\description{
Estimate treatment probability
}
