% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VignetteHelper.R
\name{filterAndTidyCovariatesForPs}
\alias{filterAndTidyCovariatesForPs}
\title{Preprocess cohortMethodData}
\usage{
filterAndTidyCovariatesForPs(
  cohortMethodData,
  population,
  excludeCovariateIds = c(),
  includeCovariateIds = c()
)
}
\description{
Preprocess cohortMethodData
}
