% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogistic.R
\name{marginalLogLik}
\alias{marginalLogLik}
\title{Laplace’s method for estimating marginal likelihood}
\usage{
marginalLogLik(betaMAP, sigma, p, logLik, fisherInfo)
}
\description{
Laplace’s method for estimating marginal likelihood
}
