% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotS}
\alias{plotS}
\title{Plot survival curves}
\usage{
plotS(
  S1,
  S0,
  Time = seq(0, 70, by = 1),
  freq.time = 90,
  Label = c("Thiazide-like diuretics", "ACE inhibitor"),
  Main = "Cross-fitted TMLE, S(t)"
)
}
\arguments{
\item{freq.time}{Map time interval to coarser intervals}

\item{Label}{Treatment labels (comparator, treatment)}

\item{Main}{Title of the plot}
}
\description{
Plot survival curves
}
