% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogistic.R
\name{pooled_design_iter}
\alias{pooled_design_iter}
\title{Calculate X^T mu, D and X^T diag(D) X for pooled logistic regression}
\usage{
pooled_design_iter(
  X_baseline,
  temporal_effect,
  timeEffect,
  beta,
  indx_subset,
  maxTime
)
}
\arguments{
\item{X_baseline}{Baseline variables that won't interact with time in regression,
sparse matrix of class "dgTMatrix",
reorder observations into decreasing survival time,
intercept included}

\item{temporal_effect}{Baseline variables that will interact with time in regression,
sparse matrix of class "dgTMatrix" or matrix,
reorder observations into decreasing survival time,
intercept included}

\item{timeEffect}{Functions of time in the discrete censoring hazards model.
Options currently include "linear", "cubic", NULL}

\item{beta}{Current iteration of coefficient value}

\item{indx_subset}{Subset index for each time point}
}
\value{
A list
}
\description{
Calculate X^T mu, D and X^T diag(D) X for pooled logistic regression
}
