% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogistic.R
\name{pooled_design_matvec}
\alias{pooled_design_matvec}
\title{Calculate X^T y for pooled logistic regression (this part doesn't need to iterate)}
\usage{
pooled_design_matvec(
  X_baseline,
  temporal_effect,
  weight,
  timeEffect,
  Y,
  timeIntMidPoint,
  indx_subset,
  maxTime,
  nsBase
)
}
\arguments{
\item{X_baseline}{Baseline variables that won't interact with time in regression,
sparse matrix of class "dgTMatrix",
reorder observations into decreasing survival time,
Intercept included.}

\item{temporal_effect}{Baseline variables that will interact with time in regression,
sparse matrix of class "dgTMatrix" or matrix,
reorder observations into decreasing survival time,
intercept included if there is temporal effect}

\item{timeEffect}{Functions of time in the discrete censoring hazards model.
Options currently include "linear", "cubic", NULL}

\item{Y}{Outcome variable in the pooled logistic regression.
Long-format, include outcome with all individuals at each time point}

\item{indx_subset}{Subset index for each time point}

\item{maxTime}{Maximum time for estimation}
}
\value{
A vector
}
\description{
Calculate X^T y for pooled logistic regression (this part doesn't need to iterate)
}
