% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogistic.R
\name{predict_pooled}
\alias{predict_pooled}
\title{Prediction for pooled logistic regression}
\usage{
predict_pooled(coef, X_baseline, temporal_effect, timeEffect, maxTime)
}
\arguments{
\item{X_baseline}{Baseline variables that won't interact with time in regression,
sparse matrix of class "dgTMatrix".
Need to include intercept}

\item{temporal_effect}{Baseline variables that will interact with time in regression,
sparse matrix of class "dgTMatrix" or matrix.
Need to include intercept if is.temporal = TRUE}

\item{timeEffect}{Functions of time in the discrete censoring hazards model.
Options currently include "linear", "cubic", NULL}
}
\value{
Probability, in the order of: id1(t1, t2....), id2(t1, t2....)....
}
\description{
Prediction for pooled logistic regression
}
