% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{simCenTime}
\alias{simCenTime}
\title{Simulate censoring time, nonproportional}
\usage{
simCenTime(treatment, covariates, lambda1, nu1, lambda0, nu0, seed)
}
\description{
Simulate censoring time, nonproportional
}
