% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{simulateData}
\alias{simulateData}
\title{Simulate time-to-event data with survival and censoring hazards}
\usage{
simulateData(treatment, survHaz, cenHaz, coarsenedTime, seed)
}
\arguments{
\item{survHaz}{Output from estimateHaz}

\item{cenHaz}{Output from estimateHaz}

\item{nInt}{number of time intervals for coarsening the data}
}
\description{
Simulate time-to-event data with survival and censoring hazards
}
