% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedCox.R
\name{strataCox}
\alias{strataCox}
\title{Estimate stratified cox model of survival probability / rmst}
\usage{
strataCox(
  treatment,
  eventObserved,
  time,
  treatProb,
  cenHaz,
  timeIntMidPoint,
  timeIntLength,
  breaks,
  nsim,
  printSim,
  hazMethod
)
}
\arguments{
\item{treatProb}{Estimated probability for treatment for each person if receive treatment 1}

\item{cenHaz}{}
}
\description{
Estimate stratified cox model of survival probability / rmst
}
