% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW.R
\name{estimateAIPWprob}
\alias{estimateAIPWprob}
\title{Estimate (cross-fitted) Augmented IPW of survival probability / rmst at time tau}
\usage{
estimateAIPWprob(
  treatment,
  eventObserved,
  time,
  survHaz,
  cenHaz,
  treatProb,
  tau,
  estimand,
  printTau
)
}
\arguments{
\item{survHaz}{Data frame with two columns: Haz1, Haz0
Estimated survival hazards for each person at each time points if receive treatment 1 (SurvHaz1) and if receive treatment 0 (SurvHaz0)}

\item{cenHaz}{Data frame with two columns: Haz1, Haz0
Estimated censoring hazards for each person at each time points if receive treatment 1 (CenHaz1) and if receive treatment 0 (CenHaz0)}

\item{treatProb}{Estimated probability for treatment for each person if receive treatment 1}

\item{tau}{Time of interest. Can be a vector (multiple time of interest)}
}
\value{
A data frame with three columns: estimand1, estimand0, SE
}
\description{
Estimate (cross-fitted) Augmented IPW of survival probability / rmst at time tau
}
