% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledLogistic.R
\name{outcomeY}
\alias{outcomeY}
\title{Output outcome Y in the long-format, include outcome with all individuals at each time point}
\usage{
outcomeY(time, eventObserved, estimate_hazard, timeIntMidPoint, maxTime)
}
\arguments{
\item{time}{Observed survival time. Ordered into decreasing observed survival time}

\item{eventObserved}{Event indicator. Ordered into decreasing observed survival time}

\item{estimate_hazard}{"survival" or "censoring"}

\item{maxTime}{Maximum time for estimation}
}
\value{
A vector
}
\description{
Output outcome Y in the long-format, include outcome with all individuals at each time point
}
